# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for work with tree comments",
        "title": "Comment Tree Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "RidusM",
            "email": "stormkillpeople@gmail.com"
        },
        "license": {
            "name": "MIT-0",
            "url": "https://github.com/aws/mit-0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/comments": {
            "get": {
                "description": "Returns the comment tree: either the roots (if parent_id is not specified) or a subtree from the specified parent. Supports pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Get the comment tree",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Parent comment ID (to get a subtree)",
                        "name": "parent_id",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment list with tree",
                        "schema": {
                            "$ref": "#/definitions/handler.CommentListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID or pagination parameters format",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Comment not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new comment with an optional parent. Returns the created object with the calculated path and depth.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Create a comment",
                "parameters": [
                    {
                        "description": "Comment data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Comment created",
                        "schema": {
                            "$ref": "#/definitions/handler.CommentResponse"
                        }
                    },
                    "400": {
                        "description": "Input validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Parent comment not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Maximum nesting depth exceeded",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/comments/search": {
            "get": {
                "description": "Full-text search by author and comment content. Returns a flat list of matches with pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Search comments",
                "parameters": [
                    {
                        "minLength": 1,
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "schema": {
                            "$ref": "#/definitions/handler.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Empty or invalid search query",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/comments/{id}": {
            "delete": {
                "description": "Soft delete comment and all its descendants (cascade). The comment is marked as is_deleted=true",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Delete comment",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment deleted",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Comment not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Return service status and current timestamp. No authentication required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Health check endpoint",
                "responses": {
                    "200": {
                        "description": "Service is healthy",
                        "schema": {
                            "$ref": "#/definitions/handler.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.CommentListResponse": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.CommentTreeResponse"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "total_count": {
                    "type": "integer",
                    "example": 800
                },
                "total_pages": {
                    "type": "integer",
                    "example": 40
                }
            }
        },
        "handler.CommentResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string",
                    "example": "admin"
                },
                "content": {
                    "type": "string",
                    "example": "hello world"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-10-27T10:00:00Z"
                },
                "depth": {
                    "type": "integer",
                    "example": 2
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "is_deleted": {
                    "type": "boolean",
                    "example": false
                },
                "parent_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                }
            }
        },
        "handler.CommentTreeResponse": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.CommentTreeResponse"
                    }
                },
                "comment": {
                    "$ref": "#/definitions/handler.CommentResponse"
                }
            }
        },
        "handler.CreateCommentRequest": {
            "type": "object",
            "required": [
                "author",
                "content"
            ],
            "properties": {
                "author": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "admin"
                },
                "content": {
                    "type": "string",
                    "maxLength": 5000,
                    "minLength": 1,
                    "example": "hello world"
                },
                "parent_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "details": {
                    "type": "string",
                    "example": "comment with id 123 does not exist"
                },
                "error": {
                    "type": "string",
                    "example": "comment not found"
                }
            }
        },
        "handler.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                },
                "time": {
                    "type": "string",
                    "example": "2026-05-08T06:04:15Z"
                }
            }
        },
        "handler.SearchResponse": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.CommentResponse"
                    }
                },
                "query": {
                    "type": "string",
                    "example": "admin"
                },
                "total_count": {
                    "type": "integer",
                    "example": 54
                }
            }
        },
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        }
    }
}